# openapi.json
{
  "components": {
    "schemas": {
...
      "Counts": {
        "properties": {
          "cas": {
            "default": 0,
            "title": "Cas",
            "type": "integer"
          },
          "health": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "ok",
                "degraded",
                "down",
                "unknown"
              ]
            },
            "title": "Health",
            "type": "object"
          },
          "links": {
            "default": 0,
            "title": "Links",
            "type": "integer"
          },
          "nodes": {
            "default": 0,
            "title": "Nodes",
            "type": "integer"
          },
          "routers": {
            "default": 0,
            "title": "Routers",
            "type": "integer"
          },
          "switches": {
            "default": 0,
            "title": "Switches",
            "type": "integer"
          }
        },
        "title": "Counts",
        "type": "object"
      },
      "Diff": {
        "description": "The union of both endpoints' topologies, annotated.\n\nUnion rather than a change-list: a removed link is absent from the `to`\ntopology, so a client given only the changes could not draw the ghost.\n\n`churn` counts events regardless of net change, because a link that flapped\nrepeatedly and came back is byte-identical in a net diff and is probably the\nmost interesting thing in the range.\n\n`system_groups` is the union of both endpoints', so a chassis that lost its\nlast node is still there for the removed children to sit inside.",
        "properties": {
          "fabric": {
            "title": "Fabric",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/DiffLink"
            },
            "title": "Links",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/DiffNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "since": {
            "$ref": "#/components/schemas/Resolved"
          },
          "system_groups": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SystemGroup"
            },
            "title": "System Groups",
            "type": "object"
          },
          "unchanged_guaranteed": {
            "default": false,
            "title": "Unchanged Guaranteed",
            "type": "boolean"
          },
          "until": {
            "$ref": "#/components/schemas/Resolved"
          }
        },
        "required": [
          "fabric",
          "since",
          "until",
          "nodes",
          "links"
        ],
        "title": "Diff",
        "type": "object"
      },
      "DiffLink": {
        "properties": {
          "after": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LinkElement"
              },
              {
                "type": "null"
              }
            ]
          },
          "before": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LinkElement"
              },
              {
                "type": "null"
              }
            ]
          },
          "change": {
            "enum": [
              "added",
              "removed",
              "modified",
              "unchanged"
            ],
            "title": "Change",
            "type": "string"
          },
          "churn": {
            "default": 0,
            "title": "Churn",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "change"
        ],
        "title": "DiffLink",
        "type": "object"
      },
      "DiffNode": {
        "properties": {
          "after": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeElement"
              },
              {
                "type": "null"
              }
            ]
          },
          "before": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeElement"
              },
              {
                "type": "null"
              }
            ]
          },
          "change": {
            "enum": [
              "added",
              "removed",
              "modified",
              "unchanged"
            ],
            "title": "Change",
            "type": "string"
          },
          "churn": {
            "default": 0,
            "title": "Churn",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "change"
        ],
        "title": "DiffNode",
        "type": "object"
      },
...
      "Event": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "event_id": {
            "title": "Event Id",
            "type": "integer"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "link_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Id"
          },
          "node_guid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Guid"
          },
          "node_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Label"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "peer_guid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peer Guid"
          },
          "peer_port_num": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peer Port Num"
          },
          "port_num": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port Num"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          }
        },
        "required": [
          "event_id",
          "occurred_at",
          "event_type"
        ],
        "title": "Event",
        "type": "object"
      },
      "Events": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "title": "Events",
            "type": "array"
          },
          "fabric": {
            "title": "Fabric",
            "type": "string"
          },
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          },
          "until": {
            "format": "date-time",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "fabric",
          "since",
          "until",
          "events"
        ],
        "title": "Events",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinkElement": {
        "properties": {
          "health": {
            "enum": [
              "ok",
              "degraded",
              "down",
              "unknown"
            ],
            "title": "Health",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "rate_gbps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Gbps"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_port": {
            "title": "Source Port",
            "type": "integer"
          },
          "speed": {
            "$ref": "#/components/schemas/Rate"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "target_port": {
            "title": "Target Port",
            "type": "integer"
          },
          "width": {
            "$ref": "#/components/schemas/Rate"
          }
        },
        "required": [
          "id",
          "source",
          "target",
          "source_port",
          "target_port",
          "health"
        ],
        "title": "LinkElement",
        "type": "object"
      },
...
      "NodeElement": {
        "properties": {
          "health": {
            "enum": [
              "ok",
              "degraded",
              "down",
              "unknown"
            ],
            "title": "Health",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "num_ports": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Ports"
          },
          "sm_role": {
            "anyOf": [
              {
                "enum": [
                  "discovering",
                  "standby",
                  "master"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sm Role"
          },
          "system_image_guid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Image Guid"
          },
          "type": {
            "enum": [
              "ca",
              "switch",
              "router"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "type",
          "health"
        ],
        "title": "NodeElement",
        "type": "object"
      },
...
      "Rate": {
        "description": "A speed or width, as stored and as read.",
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "mask": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mask"
          }
        },
        "title": "Rate",
        "type": "object"
      },
      "Resolved": {
        "description": "Which sweep answered this request. Present on every time-parameterised\nresponse.\n\nBuilt from the `db.Resolved` dataclass the pool hands back.",
        "properties": {
          "collected_at": {
            "format": "date-time",
            "title": "Collected At",
            "type": "string"
          },
          "latest_sweep_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Sweep At"
          },
          "latest_sweep_complete": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Sweep Complete"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "integer"
          },
          "topology_hash": {
            "title": "Topology Hash",
            "type": "string"
          }
        },
        "required": [
          "requested_at",
          "snapshot_id",
          "collected_at",
          "topology_hash"
        ],
        "title": "Resolved",
        "type": "object"
      },
...
      "SystemGroup": {
        "properties": {
          "children": {
            "items": {
              "type": "string"
            },
            "title": "Children",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "children"
        ],
        "title": "SystemGroup",
        "type": "object"
      },
      "Topology": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/Counts"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkElement"
            },
            "title": "Links",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeElement"
            },
            "title": "Nodes",
            "type": "array"
          },
          "resolved": {
            "$ref": "#/components/schemas/Resolved"
          },
          "system_groups": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SystemGroup"
            },
            "title": "System Groups",
            "type": "object"
          }
        },
        "required": [
          "resolved",
          "counts",
          "nodes",
          "links",
          "system_groups"
        ],
        "title": "Topology",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Read-only view of the InfiniBand fabric, at any point in time.",
    "title": "ibmon read API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/fabrics/{fabric}/diff": {
      "get": {
        "description": "The union of both topologies, each element tagged with how it changed.",
        "operationId": "diff_api_v1_fabrics__fabric__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "fabric",
            "required": true,
            "schema": {
              "title": "Fabric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diff"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff",
        "tags": [
          "fabric"
        ]
      }
    },
    "/api/v1/fabrics/{fabric}/events": {
      "get": {
        "description": "Newest first. Defaults to the last 24 hours.\n\nComplements the diff rather than duplicating it: a diff says a link is\nunchanged, this says it went down and came back forty times while nobody\nwas looking.",
        "operationId": "events_api_v1_fabrics__fabric__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "fabric",
            "required": true,
            "schema": {
              "title": "Fabric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Events"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "fabric"
        ]
      }
    },
...
    "/api/v1/fabrics/{fabric}/topology": {
      "get": {
        "description": "The graph as of `at`, or now.",
        "operationId": "topology_api_v1_fabrics__fabric__topology_get",
        "parameters": [
          {
            "in": "path",
            "name": "fabric",
            "required": true,
            "schema": {
              "title": "Fabric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topology"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Topology",
        "tags": [
          "fabric"
        ]
      }
    }
  }
}